% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_seconds.R
\name{convert_to_secs}
\alias{convert_to_secs}
\title{Converts Times To Seconds}
\usage{
convert_to_secs(times)
}
\arguments{
\item{times}{character vector of times}
}
\description{
Utility function that pads times with the appropriate number of copies
of "00:" to put it in hh:mm:ss.s format and then converts to seconds.
}
