% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_clean.R
\name{dst_clean_opa}
\alias{dst_clean_opa}
\title{Alternate version for the PDF format typically seen in OPA Cups}
\usage{
dst_clean_opa(tbls, race_distance)
}
\description{
Alternate version for the PDF format typically seen in OPA Cups
}
