% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_plot.R
\name{dst_move_plot}
\alias{dst_move_plot}
\title{Plot Distance Races Split Movement}
\usage{
dst_move_plot(data, n = 5, top_n = Inf, offset.x = 0.25)
}
\arguments{
\item{data}{data.frame; as returned by \code{\link{parse_dst_pdf}}}

\item{n}{integer; number of athletes to highlight in each direction}

\item{top_n}{integer; focus on athletes moving into or out of the top_n}

\item{offset.x}{numeric; amount to slide the names to the left/right}
}
\description{
Highlight movement up and down by skiers throughout a distance race.
}
