% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_pdf}
\alias{parse_pdf}
\title{Parse PDF File}
\usage{
parse_pdf(file, ...)
}
\arguments{
\item{file}{character; file path}

\item{\dots}{further arguments to \code{\link[tabulizer]{extract_tables}}}
}
\description{
Wrapper for \code{\link[tabulizer]{extract_tables}}
}
