% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_parser.R
\name{parse_spr_pdf}
\alias{parse_spr_pdf}
\title{Parse Sprint Race PDFs}
\usage{
parse_spr_pdf(final_pdf = NULL)
}
\arguments{
\item{final_pdf}{file path to PDF of final sprint results}
}
\description{
Convert FIS sprint result PDFs into a format more
suitable for analysis. All times are converted to seconds.
}
\examples{
\dontrun{
spr <- parse_spr_df(final_pdf = system.file("inst/example_pdfs/sprint_final.pdf",package = "fispdfparsr"))
}
}

