% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_parser.R
\name{parse_stage_pdf}
\alias{parse_stage_pdf}
\title{Parse Stage Race PDFs}
\usage{
parse_stage_pdf(file)
}
\description{
Parses results from stage race standings PDFs, i.e. the PDFs with cumulative
results from multiple stages.
}
\examples{
\dontrun{
stg <- parse_stage_pdf(file = system.file("example_pdfs/stage_example1.pdf",
                                           package = "fispdfparsr"))
}
}

