% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_wc_pts_pdf.R
\name{parse_wc_pts_pdf}
\alias{parse_wc_pts_pdf}
\title{Parse World Cup Point PDFs}
\usage{
parse_wc_pts_pdf(file)
}
\arguments{
\item{file}{character; file path to PDF}
}
\description{
Parse World Cup Point PDFs
}
\examples{
\dontrun{
stg <- parse_wc_pts_pdf(file = system.file("example_pdfs/wc_pts_example1.pdf",package = "fispdfparsr"))
}
}
