% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_plot.R
\name{stage_plot}
\alias{stage_plot}
\title{Stage Results Plot}
\usage{
stage_plot(
  data,
  type = c("rank", "time", "percent"),
  nation_col = NULL,
  name_col = NULL,
  offset.x = 0.05
)
}
\arguments{
\item{data}{data.frame; output from \code{\link{parse_stage_pdf}}}

\item{type}{character; one of "rank", "time" or "percent" to plot
overall rank, time back or percent back.}

\item{nation_col}{character vector of nations to highlight}

\item{name_col}{character vector of athletes to highlight}

\item{offset.x}{numeric; amount to adjust the athlete labels on the
left and right of the plot}
}
\description{
Plot bumps chart for stage race results.
}
\examples{
\dontrun{
pdf <- system.file("example_pdfs/stage_example2.pdf",package = "fispdfparsr")
stg <- parse_stage_pdf(file = pdf)
require(ggplot2)
p <- stage_plot(data = stg,type = "rank")
print(p)
}
}
