% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_parser.R
\name{parse_dst_pdf}
\alias{parse_dst_pdf}
\title{Parse Distance Race PDFs}
\usage{
parse_dst_pdf(
  file = NULL,
  race_distance,
  long_mass = FALSE,
  edit = FALSE,
  opa = FALSE,
  ...
)
}
\arguments{
\item{file}{file path to PDF of distance results}

\item{race_distance}{numeric; race distance in km}

\item{long_mass}{boolean; flag for handling long mass start start races}

\item{edit}{boolean}
}
\description{
Convert FIS distance result PDFs into a format more
suitable for analysis. All times are converted to seconds.
}
\examples{
\dontrun{
dst <- parse_dst_pdf(file = system.file("example_pdfs/dst_example1.pdf",
                                        package = "fispdfparsr"),15)
}
}
