% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_parser.R
\name{parse_dst_pdf}
\alias{parse_dst_pdf}
\title{Parse Distance Race PDFs}
\usage{
parse_dst_pdf(race_pdf = NULL, race_distance = NULL)
}
\arguments{
\item{race_pdf}{file path to PDF of distance results}

\item{race_distance}{numeric; race distance in km}
}
\description{
Convert FIS distance result PDFs into a format more
suitable for analysis. All times are converted to seconds.
}
\examples{
\dontrun{
dst <- parse_dst_df(final_pdf = system.file("inst/example_pdfs/men_dst.pdf",package = "fispdfparsr"))
}
}

