% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_parser.R
\name{parse_spr_pdf}
\alias{parse_spr_pdf}
\title{Parse Sprint Race PDFs}
\usage{
parse_spr_pdf(file = NULL)
}
\arguments{
\item{file}{file path to PDF of final sprint results}
}
\description{
Convert FIS sprint result PDFs into a format more
suitable for analysis. All times are converted to seconds.
}
\details{
The PDF parser in the tabulizer package will sometimes miss blank columns,
for example if no one from QF3 ended up on the final heat, the parser will
simply skip that blank column. In these cases you will be prompted to supply
an index for where to insert a blank column so that the tables all line up
correctly. You might need to have the PDF open to look at while running this
function so you can verify where to add the column.
}
\examples{
\dontrun{
spr <- parse_spr_df(file = system.file("inst/example_pdfs/spr_example1.pdf",package = "fispdfparsr"))
}
}

